==> docs/lesson5/5_assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd
import numpy as np

df08 = pd.read_csv('all_alpha_08.csv')
df08.head()

df18 = pd.read_csv('all_alpha_18.csv')
df18.head()

df08.shape

sum(df08.duplicated())

# trying to count rows with missing data. these produce the same result.
# np.count_nonzero(df08.isnull())
# sum(df08.isnull().sum())

# but taking the max of this is apparently correct
df08.isnull().sum()

df08.info()

df08.nunique()

df18.shape

# taking the max of this is apparently correct for rows w/ missing data
df18.isnull().sum()

df18.info()

df18.nunique()

sum(df18.duplicated())

type(df08['Cyl'][0])

type(df18['Cyl'][0])

type(df18['City MPG'][0])

type(df08['Greenhouse Gas Score'][0])

type(df18['Greenhouse Gas Score'][0])

df08.Fuel.unique()

df18.Fuel.unique()
