==> docs/lesson5/8_inspect_datatypes.ipynb <==
# Inspecting Data Types
Use the space below to explore `data_08.csv` and `data_18.csv` to answer the quiz questions below regarding datatypes.

import pandas as pd

df_08 = pd.read_csv('data_08.csv')
df_08.head(1)

df_18 = pd.read_csv('data_18.csv')
df_18.head(1)

df_08['cyl'].unique()

df_18['cyl'].unique()

df_08['air_pollution_score'].unique()

df_18['air_pollution_score'].unique()

type(df_08['city_mpg'][0])
type(df_18['city_mpg'][0])

df_08['greenhouse_gas_score'].unique()

df_18['greenhouse_gas_score'].unique()
