==> docs/lesson5/11_fix_datatypes_mpg_greenhouse.ipynb <==
## Fix `city_mpg`, `hwy_mpg`, `cmb_mpg` datatypes
    2008 and 2018: convert string to float

# load datasets
import pandas as pd

df_08 = pd.read_csv('data_08.csv')
df_08.head(1)

df_18 = pd.read_csv('data_18.csv')
df_18.head(1)

# convert mpg columns to floats
mpg_columns = ['city_mpg', 'hwy_mpg', 'cmb_mpg']
for c in mpg_columns:
    df_18[c] = df_18[c].astype(float)
    df_08[c] = df_08[c].astype(float)

print('s/b float: {}'.format(type(df_08['city_mpg'][0])))
print('s/b float: {}'.format(type(df_18['city_mpg'][0])))
print('s/b int: {}'.format(type(df_08['cyl'][0])))
print('s/b int: {}'.format(type(df_18['cyl'][0])))
print('s/b float: {}'.format(type(df_08['air_pollution_score'][0])))
print('s/b float: {}'.format(type(df_18['air_pollution_score'][0])))
print('s/b int: {}'.format(type(df_08['greenhouse_gas_score'][0])))
print('s/b int: {}'.format(type(df_18['greenhouse_gas_score'][0])))

## Fix `greenhouse_gas_score` datatype
    2008: convert from float to int

# convert from float to int
df_08['greenhouse_gas_score'] = df_08['greenhouse_gas_score'].astype(int)

## All the dataypes are now fixed! Take one last check to confirm all the changes.

df_08.dtypes

df_18.dtypes

df_08.dtypes == df_18.dtypes

# Save your new CLEAN datasets as new files!
df_08.to_csv('clean_08.csv', index=False)
df_18.to_csv('clean_18.csv', index=False)
