==> docs/lesson5/15_merging_data.ipynb <==
# Merging Datasets
Use Pandas Merges to create a combined dataset from `clean_08.csv` and `clean_18.csv`.

# load datasets
import pandas as pd

df_08 = pd.read_csv('clean_08.csv')
df_08.head(1)

df_18 = pd.read_csv('clean_18.csv')
df_18.head(1)

### Create combined dataset

# rename 2008 columns
df_08.rename(columns=lambda x: x[:10] + "_2008", inplace=True)

# view to check names
df_08.head()

# merge datasets
df_combined = pd.merge(df_08, df_18, left_on='model_2008', right_on='model')

# view to check merge
df_combined.head()

df_combined.shape

Save the combined dataset

df_combined.to_csv('combined_dataset.csv', index=False)
